Successfuly Seeded but Relations not working

+++ Models/BooksAuthors.cs
@@ -9,9 +9,9 @@ namespace LibraryAPI.Models
     public class BooksAuthors
     {
         [ForeignKey("BookId")]
-        public int BookId { get; set; }
+        public long BookId { get; set; }
         [ForeignKey("AuthorId")]
-        public int AuthorId { get; set; }
+        public long AuthorId { get; set; }
         public Book Book { get; set; }
         public Author Author { get; set; }
     }
