+++ Models/BooksGenres.cs
@@ -9,9 +9,9 @@ namespace LibraryAPI.Models
     public class BooksGenres
     {
         [ForeignKey("BookId")]
-        public int BookId { get; set; }
+        public long BookId { get; set; }
         [ForeignKey("GenreId")]
-        public int GenreId { get; set; }
+        public long GenreId { get; set; }
         public Book Book { get; set; }
         public Genre Genre { get; set; }
     }
